+++ src/test/java/io/anserini/collection/DocumentTest.java
@@ -14,18 +14,17 @@
  * limitations under the License.
  */
 
-package io.anserini.document;
+package io.anserini.collection;
+
+import org.apache.lucene.util.LuceneTestCase;
+import org.junit.Before;
+import org.junit.Test;
 
 import java.io.BufferedReader;
 import java.io.StringReader;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-
-import io.anserini.collection.SourceDocument;
-import org.apache.lucene.util.LuceneTestCase;
-import org.junit.Before;
-import org.junit.Test;
 
 public class DocumentTest<D extends SourceDocument> extends LuceneTestCase {
   protected List<String> rawDocs;
@@ -35,8 +34,8 @@ public class DocumentTest<D extends SourceDocument> extends LuceneTestCase {
   @Before
   public void setUp() throws Exception {
     super.setUp();
-    rawDocs = new ArrayList<String>();
-    expected = new ArrayList<Map<String, String>>();
+    rawDocs = new ArrayList<>();
+    expected = new ArrayList<>();
   }
 
   protected D parse(String raw) throws Exception {
