+++ src/test/java/io/anserini/collection/TrecDocumentTest.java
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package io.anserini.document;
+package io.anserini.collection;
 
-import java.util.HashMap;
-
+import io.anserini.collection.DocumentTest;
 import io.anserini.collection.TrecCollection;
 import org.junit.Before;
+
+import java.util.HashMap;
 
 public class TrecDocumentTest extends DocumentTest<TrecCollection.Document> {
